Fetch GitHub project board at build time so the token never ships to the browser
