+++ src/data/githubProjectBoard.json
@@ -1,0 +1,4 @@
+{
+  "items": [],
+  "error": null
+}
